+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - #101 - Add fuzzy search for product
 - #82 - Generate test data for PickupPointRestController
 - #63 - Front CRUD for PickupPoint
+- #105 - Status model for User
 
 ## [1.1.0] - 8.09.2023
 - #5 - added user entity